123
123

+++ src/_02_diy_array_list/CustomArrayListTest.java
@@ -13,9 +13,9 @@ public class CustomArrayListTest {
 		list.add(1);
 		list.add(206);
 	
-		if(list.get(0) != 0) fail("list.get(0) returning improper value.");
-		if(list.get(1) != 1) fail("list.get(1) returning improper value.");
-		if(list.get(2) != 206) fail("list.get(2) returning improper value.");
+		if((int)list.get(0) != 0) fail("list.get(0) returning improper value.");
+		if((int)list.get(1) != 1) fail("list.get(1) returning improper value.");
+		if((int)list.get(2) != 206) fail("list.get(2) returning improper value.");
 	}
 	
 	@Test
